Move config to internal package
